requirements.txt
Django==1.5.5
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==1.6
Sphinx==1.2.2
argparse==1.2.1
bleach==1.4
django-braces==1.4.0
django-debug-toolbar==0.9.4
docutils==0.12
html5lib==0.999
mock==1.0.1
model-mommy==1.2.1
six==1.3.0
wsgiref==0.1.2
